refactor(edge): move http/client to edge/client
